Minor improvements

+++ src/main/java/io/p1jmonitor/p1processor/read/CRC16.java
@@ -51,7 +51,7 @@ public class CRC16 implements Checksum {
 
     public void update(byte[] b, int off, int len) {
         for (int i = off; i < off+len; i++)
-            update((int)b[i]);
+            update(b[i]);
     }
 
     public void update(int b) {
